+++ src/main/java/com/ilkun/delivery/repository/InMemBonusCardRepository.java
@@ -1,6 +1,8 @@
 package com.ilkun.delivery.repository;
 
 import com.ilkun.delivery.domain.BonusCard;
+import com.ilkun.delivery.repository.BonusCardRepository;
+import com.ilkun.delivery.repository.BonusCardRepository;
 import java.util.HashMap;
 import java.util.Map;
 import javax.annotation.PostConstruct;
@@ -9,8 +11,8 @@ import org.springframework.stereotype.Repository;
 @Repository
 public class InMemBonusCardRepository implements BonusCardRepository {
 
-    private static int curId = 1;
-    private static final Map<Integer, BonusCard> bonusCards = new HashMap<>();
+    private static long curId = 1;
+    private static final Map<Long, BonusCard> bonusCards = new HashMap<>();
 
     @PostConstruct
     public void init() {
@@ -18,7 +20,7 @@ public class InMemBonusCardRepository implements BonusCardRepository {
     }
     
     @Override
-    public BonusCard find(Integer id) {
+    public BonusCard find(Long id) {
         return bonusCards.get(id);
     }
